requirements.txt
cvxpy==1.4.2
matplotlib==3.8.0
numpy==1.24.3
python_papi==5.5.1.6
pytorch_lightning==2.0.2
PyYAML==6.0
seaborn==0.13.2
torch==2.0.0
torch_geometric==2.3.0
